// src/app/generated/graphql.ts
export class AllAddressesByCustomerGQL extends Apollo.Query<
  AllAddressesByCustomer.Query,
  AllAddressesByCustomer.Variables
> {
  document: any = gql`
    query allAddressesByCustomer($customerId: UUID!) {
      allAddresses(filter: { customerId: { equalTo: $customerId } }) {
        totalCount
        nodes {
          id
          type
          name
          firstName
          lastName
          company
          line1
          line2
          city
          postcode
          country
          instructions
          defaultAddress
        }
      }
    }
  `;
}
// ...
export class CreateAddressGQL extends Apollo.Mutation<
  CreateAddress.Mutation,
  CreateAddress.Variables
> {
  document: any = gql`
    mutation createAddress(
      $customerId: UUID!
      $type: AddressType!
      $name: String
      $firstName: String!
      $lastName: String!
      $company: String
      $line1: String!
      $line2: String
      $city: String!
      $postcode: String!
      $country: String!
      $instructions: String
      $defaultAddress: Boolean
    ) {
      createAddress(
        input: {
          address: {
            customerId: $customerId
            type: $type
            name: $name
            firstName: $firstName
            lastName: $lastName
            company: $company
            line1: $line1
            line2: $line2
            city: $city
            postcode: $postcode
            country: $country
            instructions: $instructions
            defaultAddress: $defaultAddress
          }
        }
      ) {
        address {
          id
        }
      }
    }
  `;
}
// ...
export type Maybe<T> = T | null;
// ...
export enum AddressType {
  Billing = "BILLING",
  Shipping = "SHIPPING"
}
// ...
export type Uuid = any;
// ...
export namespace AllAddressesByCustomer {
  export type Variables = {
    customerId: Uuid;
  };

  export type Query = {
    __typename?: "Query";

    allAddresses: Maybe<AllAddresses>;
  };

  export type AllAddresses = {
    __typename?: "AddressesConnection";

    totalCount: Maybe<number>;

    nodes: (Maybe<Nodes>)[];
  };

  export type Nodes = {
    __typename?: "Address";

    id: Uuid;

    type: AddressType;

    name: Maybe<string>;

    firstName: string;

    lastName: string;

    company: Maybe<string>;

    line1: string;

    line2: Maybe<string>;

    city: string;

    postcode: string;

    country: string;

    instructions: Maybe<string>;

    defaultAddress: boolean;
  };
}

export namespace AuthenticateUserCustomer {
  export type Variables = {
    email: string;
    password: string;
  };

  export type Mutation = {
    __typename?: "Mutation";

    authenticateUserCustomer: Maybe<AuthenticateUserCustomer>;
  };

  export type AuthenticateUserCustomer = {
    __typename?: "AuthenticateUserCustomerPayload";

    jwtToken: Maybe<JwtToken>;
  };
}
// ...
export namespace CreateAddress {
  export type Variables = {
    customerId: Uuid;
    type: AddressType;
    name?: Maybe<string>;
    firstName: string;
    lastName: string;
    company?: Maybe<string>;
    line1: string;
    line2?: Maybe<string>;
    city: string;
    postcode: string;
    country: string;
    instructions?: Maybe<string>;
    defaultAddress?: Maybe<boolean>;
  };

  export type Mutation = {
    __typename?: "Mutation";

    createAddress: Maybe<CreateAddress>;
  };

  export type CreateAddress = {
    __typename?: "CreateAddressPayload";

    address: Maybe<Address>;
  };

  export type Address = {
    __typename?: "Address";

    id: Uuid;
  };
}